[코리] Step-6 댓글 구현

+++ src/main/java/codesquad/springcafe/service/ArticleService.java
@@ -2,21 +2,27 @@ package codesquad.springcafe.service;
 
 import codesquad.springcafe.dto.ArticleRequestDto;
 import codesquad.springcafe.exception.ArticleNotFountException;
+import codesquad.springcafe.exception.CannotDeleteArticleWithRepliesException;
+import codesquad.springcafe.exception.UnauthorizedAccessException;
 import codesquad.springcafe.model.Article;
+import codesquad.springcafe.model.Reply;
 import codesquad.springcafe.repository.ArticleRepository;
+import codesquad.springcafe.repository.ReplyRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
 
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, ReplyRepository replyRepository) {
         this.articleRepository = articleRepository;
+        this.replyRepository = replyRepository;
     }
-
 
     public void createArticle(Article article) {
         articleRepository.save(article);
@@ -40,6 +46,17 @@ public class ArticleService {
     }
 
     public void delete(Long articleId) {
+        List<Reply> replies = replyRepository.findRepliesByArticleId(articleId);
+        if (!replies.isEmpty()) {
+            throw new CannotDeleteArticleWithRepliesException();
+        }
         articleRepository.delete(articleId);
+    }
+
+    public void validateArticleWriter(Long articleId, String userId) {
+        Article article = findById(articleId);
+        if (!article.checkWriter(userId)) {
+            throw new UnauthorizedAccessException("다른 사람의 게시글에 접근할 수 없습니다.");
+        }
     }
 }
